don't register user-excluded runtimes
