+++ extensions/positron-python/src/client/positron/interpreterSettings.ts
@@ -5,8 +5,9 @@
 
 import path from 'path';
 
-import { traceLog } from '../logging';
+import { traceInfo, traceLog, traceVerbose } from '../logging';
 import { getConfiguration } from '../common/vscodeApis/workspaceApis';
+import { arePathsSame, isParentPath } from '../pythonEnvironments/common/externalDependencies';
 
 // Settings keys for various interpreter settings
 export const INTERPRETERS_INCLUDE_SETTING_KEY = 'interpreters.include';
@@ -50,4 +51,39 @@ export function getUserExcludedInterpreters(): string[] {
     }
     traceLog(`[getUserExcludedInterpreters]: No interpreters specified via ${INTERPRETERS_EXCLUDE_SETTING_KEY}`);
     return [];
+}
+
+/**
+ * Check whether an interpreter should be included in the list of discovered interpreters.
+ * If an interpreter is both explicitly included and excluded, it will be included.
+ * @param interpreterPath The interpreter path to check
+ * @returns Whether the interpreter should be included in the list of discovered interpreters.
+ */
+export function shouldIncludeInterpreter(interpreterPath: string): boolean {
+    // If a user has explicitly included the interpreter, include it. In other words, including an
+    // interpreter takes precedence over excluding it.
+    const interpretersInclude = getUserIncludedInterpreters();
+    if (interpretersInclude.length > 0) {
+        const userIncluded = interpretersInclude.some(
+            (includePath) => isParentPath(interpreterPath, includePath) || arePathsSame(interpreterPath, includePath),
+        );
+        if (userIncluded) {
+            traceInfo(`[shouldIncludeInterpreter] Interpreter ${interpreterPath} was included via settings`);
+            return true;
+        }
+    }
+
+    // If the user has not explicitly included the interpreter, check if it is explicitly excluded.
+    const interpretersExclude = getUserExcludedInterpreters();
+    const userExcluded = interpretersExclude.some(
+        (excludePath) => isParentPath(interpreterPath, excludePath) || arePathsSame(interpreterPath, excludePath),
+    );
+    if (userExcluded) {
+        traceInfo(`[shouldIncludeInterpreter] Interpreter ${interpreterPath} was excluded via settings`);
+        return false;
+    }
+
+    // If the interpreter is not explicitly included or excluded, include it.
+    traceVerbose(`[shouldIncludeInterpreter] Interpreter ${interpreterPath} not explicitly included or excluded`);
+    return true;
 }
